requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
colorama==0.4.6
fastapi==0.115.12
h11==0.14.0
idna==3.10
psycopg==3.2.6
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
PyJWT==2.10.1
python-dotenv==1.1.0
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uvicorn==0.34.0
